+++ packages/orq-rc/src/models/operations/listdatasetdatapoints.ts
@@ -3101,7 +3101,7 @@ export const ListDatasetDatapointsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-09-30T14:52:41.173Z",
+    "2025-10-01T14:07:15.793Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -3165,7 +3165,7 @@ export const ListDatasetDatapointsData$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-09-30T14:52:41.173Z"))
+  updated: z.date().default(() => new Date("2025-10-01T14:07:15.793Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {